+++ ssvsigner/client_test.go
@@ -21,8 +21,8 @@ import (
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/zap"
 
-	"github.com/ssvlabs/ssv/ssvsigner/testingutils"
 	ssvsignertls "github.com/ssvlabs/ssv/ssvsigner/tls"
+	"github.com/ssvlabs/ssv/ssvsigner/tls/testingutils"
 	"github.com/ssvlabs/ssv/ssvsigner/web3signer"
 )
 
@@ -1205,31 +1205,25 @@ func TestTLSConfigCLIIntegration(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			// 1. Load options from ssvsignertls.ClientConfig
-			options, err := LoadTLSOptions(tc.tlsConfig)
-
 			if tc.expectError {
 				require.Error(t, err)
 				return
 			}
 
-			require.NoError(t, err)
-			require.Len(t, options, tc.expectOptions)
-
 			if tc.expectOptions == 0 {
 				return // No need to check further
 			}
 
-			// 2. Create client with options
 			logger, _ := zap.NewDevelopment()
-			allOptions := append([]ClientOption{WithLogger(logger)}, options...)
+			allOptions := append([]ClientOption{WithLogger(logger)},
+				WithClientTLSCertificates(clientCert, clientKey, caCert),
+				WithClientInsecureSkipVerify())
 
 			// Use localhost as a dummy endpoint
 			client, err := NewClient("https://localhost:8000", allOptions...)
 			require.NoError(t, err)
 			require.NotNil(t, client)
 
-			// 3. Verify TLS configuration was applied
 			transport, ok := client.httpClient.Transport.(*http.Transport)
 			require.True(t, ok)
 			require.NotNil(t, transport)
@@ -1246,150 +1240,6 @@ func TestTLSConfigCLIIntegration(t *testing.T) {
 
 			if tc.tlsConfig.ClientCertFile != "" && tc.tlsConfig.ClientKeyFile != "" {
 				require.NotEmpty(t, transport.TLSClientConfig.Certificates)
-			}
-		})
-	}
-}
-
-func TestLoadTLSOptions(t *testing.T) {
-	t.Parallel()
-
-	// Create temporary files for testing
-	clientCertFile, err := os.CreateTemp("", "client-cert-*.pem")
-	require.NoError(t, err)
-	defer os.Remove(clientCertFile.Name())
-
-	clientKeyFile, err := os.CreateTemp("", "client-key-*.pem")
-	require.NoError(t, err)
-	defer os.Remove(clientKeyFile.Name())
-
-	caCertFile, err := os.CreateTemp("", "ca-cert-*.pem")
-	require.NoError(t, err)
-	defer os.Remove(caCertFile.Name())
-
-	// Generate test certificates
-	caCert, _, clientCert, clientKey := testingutils.GenerateCertificates(t, "localhost")
-
-	// Write test data to temp files
-	err = os.WriteFile(clientCertFile.Name(), clientCert, 0644)
-	require.NoError(t, err)
-
-	err = os.WriteFile(clientKeyFile.Name(), clientKey, 0644)
-	require.NoError(t, err)
-
-	err = os.WriteFile(caCertFile.Name(), caCert, 0644)
-	require.NoError(t, err)
-
-	testCases := []struct {
-		name          string
-		config        ssvsignertls.ClientConfig
-		expectError   bool
-		expectedOpts  int
-		errorContains string
-	}{
-		{
-			name: "All certificate files",
-			config: ssvsignertls.ClientConfig{
-				ClientCertFile:   clientCertFile.Name(),
-				ClientKeyFile:    clientKeyFile.Name(),
-				ClientCACertFile: caCertFile.Name(),
-			},
-			expectError:  false,
-			expectedOpts: 1, // WithClientTLSConfig with full cert configuration
-		},
-		{
-			name: "Client cert and key only",
-			config: ssvsignertls.ClientConfig{
-				ClientCertFile: clientCertFile.Name(),
-				ClientKeyFile:  clientKeyFile.Name(),
-			},
-			expectError:  false,
-			expectedOpts: 1, // WithClientTLSConfig with client cert and key
-		},
-		{
-			name: "CA cert only",
-			config: ssvsignertls.ClientConfig{
-				ClientCACertFile: caCertFile.Name(),
-			},
-			expectError:  false,
-			expectedOpts: 1, // WithClientTLSConfig with CA cert only
-		},
-		{
-			name: "InsecureSkipVerify only",
-			config: ssvsignertls.ClientConfig{
-				ClientInsecureSkipVerify: true,
-			},
-			expectError:  false,
-			expectedOpts: 1, // WithClientTLSConfig with InsecureSkipVerify
-		},
-		{
-			name: "All options",
-			config: ssvsignertls.ClientConfig{
-				ClientCertFile:           clientCertFile.Name(),
-				ClientKeyFile:            clientKeyFile.Name(),
-				ClientCACertFile:         caCertFile.Name(),
-				ClientInsecureSkipVerify: true,
-			},
-			expectError:  false,
-			expectedOpts: 1, // WithClientTLSConfig with all options
-		},
-		{
-			name:         "No configuration",
-			config:       ssvsignertls.ClientConfig{},
-			expectError:  false,
-			expectedOpts: 0, // No options when no config provided
-		},
-		{
-			name: "Non-existent client cert file",
-			config: ssvsignertls.ClientConfig{
-				ClientCertFile: "/non/existent/file",
-				ClientKeyFile:  clientKeyFile.Name(),
-			},
-			expectError:   true,
-			errorContains: "client certificate file does not exist",
-		},
-		{
-			name: "Non-existent client key file",
-			config: ssvsignertls.ClientConfig{
-				ClientCertFile: clientCertFile.Name(),
-				ClientKeyFile:  "/non/existent/file",
-			},
-			expectError:   true,
-			errorContains: "client key file does not exist",
-		},
-		{
-			name: "Non-existent CA cert file",
-			config: ssvsignertls.ClientConfig{
-				ClientCACertFile: "/non/existent/file",
-			},
-			expectError:   true,
-			errorContains: "CA certificate file does not exist",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			opts, err := LoadTLSOptions(tc.config)
-
-			if tc.expectError {
-				require.Error(t, err)
-				if tc.errorContains != "" {
-					require.Contains(t, err.Error(), tc.errorContains)
-				}
-			} else {
-				require.NoError(t, err)
-				require.Len(t, opts, tc.expectedOpts)
-
-				if tc.expectedOpts > 0 {
-					// Create a client to verify the option works
-					logger, _ := zap.NewDevelopment()
-					allOptions := append([]ClientOption{WithLogger(logger)}, opts...)
-
-					client, err := NewClient("https://localhost:8000", allOptions...)
-					require.NoError(t, err)
-					require.NotNil(t, client)
-					require.NotNil(t, client.tlsConfig)
-				}
 			}
 		})
 	}
